requirements.txt
aniso8601==9.0.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
filelock==3.9.0
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-WTF--1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pdfkit==1.0.0
pip==23.0.1
platformdirs==3.1.1
PyPDF2==3.0.1
pytz==2022.7.1
setuptools==67.6.0
six==1.16.0
validate-email==1.3
virtualenv==20.21.0
Werkzeug==2.2.3
WTForms==3.0.1